requirements.txt
Django==3.2.6
django-cors-headers==4.2.0
djangorestframework-simplejwt
djangorestframework==3.14.0
requests
python-dotenv